+++ js.bolt.tails/src/listeners/events/link-shared.ts
@@ -3,12 +3,13 @@ import Video from "../../videos";
 
 /*
  * Cache the video at the shared link.
+ * @see https://api.slack.com/events/link_shared
  */
 export const linkShared = async ({
   event,
 }: AllMiddlewareArgs & SlackEventMiddlewareArgs<"link_shared">) => {
-  const { channel, links } = event;
-  if (channel === "COMPOSER") {
+  const { links, source } = event;
+  if (source === "composer") {
     return;
   }
   for (let n = 0; n < links.length; n++) {